(QENG-1873) Add count Section to summary Result
This adds a :count key to both the overall and daily results. The
payload is now divided in to :duration and :count values for summary
data. The daily results do not include anything under :count except the
total number of VMs cloned for the day. The overall :count includes
total, including minimum, maximum, and average VMs cloned for the date
range requested.