(maint) Remove Some Thread.new Calls
Remove the creation of threads when mostly interacting with redis. This
reduces the number of threads created during a spiking activity.

As a side effect, it is possible that a pool can take a long time
checking the port availability of new hosts. This timeout will be made
configurable in a future change.