(GH-203) Change use of abort to raise in in vsphere_helper.rb
Previously, in `find_vm` and `find_pool` if in an expected object was found the
entire pool manager, and probably API process, will terminate due to the use of
`abort`.  This commit changes the use of abort to raise so that the error can be
trapped  and handled instead of the entire process being shutdown.  This is also
required so that the methods can be tested, otherwise rspec is shutdown
prematurely.