(maint) Update rubocop.yml to ignore 2 cops
This commit updates the rubocop_todo.yml file for all of the new code that has
been introduced.

This commit adds two cops to the ignore list
- Style/ConditionalAssignment
- Next

In some cases the readability of code is better even when the cops are raised.
This commit disables the cops so they will not be flagged as violations