(maint) Pin nokogiri on legacy ruby
Due to native extensions not working for JRuby on Windows easily, sometimes very
old versions of MRI Ruby are needed for debugging (1.9.3). This commit pins
nokogiri to a version which is compatible with ruby 1.9.x.  It is expected this
is only required for debugging in edge cases. vmpooler should still be run on
modern ruby versions in production.