Update to generic launcher
I realized the launcher itself handles looking up the Ruby interpreter with /usr/bin/env ruby. Manually specifying an interpreter is going to try to circumvent that and isn't nearly as portable as an example service. By moving this sample script back to the generic launcher directly a user can choose to handle how /usr/bin/env looks up Ruby, whether it's with a jruby installation, rbenv, or however the user wishes.